+++ src/app/(provider)/(root)/deleteUser/page.tsx
@@ -1,4 +1,4 @@
-import DeleteUserInMobile from '@/components/DeleteUserInMobile';
+import DeleteUserInMobile from '@/components/mobile/DeleteUserInMobile';
 
 export default function DeleteUserPage() {
   return <DeleteUserInMobile />;
