+++ src/app/(provider)/(root)/personalInfo/page.tsx
@@ -1,4 +1,4 @@
-import DetailInfo from '@/components/DetailInfo';
+import DetailInfo from '@/components/mobile/DetailInfo';
 
 export default function PersonalInfoPage() {
   return <DetailInfo content="personal" />;
